requirements/prod.txt
APScheduler==3.10.4
Django==5.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
factory_boy==3.3.0
gunicorn==21.1.0
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-django==4.5.2
